Refactored TelemetryFileUpload to reduce complexity
